ttl now a uint64_t in storage logic

+++ storage/include/Storage.hpp
@@ -16,7 +16,7 @@ class Storage {
     Storage(const std::string& db_dir);
     ~Storage();
     bool store(const std::string& hash, const std::string& pubKey,
-               const std::string& data, int ttl);
+               const std::string& data, uint64_t ttl);
     bool retrieve(const std::string& pubKey, std::vector<storage::Item>& data,
                   const std::string& lastHash);
